ensure slot summary is applied anywhere we create agentv1.Agent
