+++ master/internal/api_agents.go
@@ -13,57 +13,8 @@ import (
 	"github.com/determined-ai/determined/master/internal/cluster"
 	"github.com/determined-ai/determined/master/internal/grpcutil"
 	"github.com/determined-ai/determined/master/internal/rm/rmerrors"
-	"github.com/determined-ai/determined/proto/pkg/agentv1"
 	"github.com/determined-ai/determined/proto/pkg/apiv1"
 )
-
-type slotStats map[string]*agentv1.DeviceStats
-
-// SummarizeSlots a set of slots.
-func SummarizeSlots(slots map[string]*agentv1.Slot) *agentv1.SlotStats {
-	stats := agentv1.SlotStats{
-		TypeStats:  make(slotStats),
-		BrandStats: make(slotStats),
-	}
-
-	if len(slots) == 0 {
-		return &stats
-	}
-	for _, slot := range slots {
-		deviceType := slot.Device.Type.String()
-		deviceTypeStats, ok := stats.TypeStats[deviceType]
-		if !ok {
-			deviceTypeStats = &agentv1.DeviceStats{
-				States: make(map[string]int32),
-			}
-			stats.TypeStats[deviceType] = deviceTypeStats
-		}
-		deviceBrand := slot.Device.Brand
-		deviceBrandStats, ok := stats.BrandStats[deviceBrand]
-		if !ok {
-			deviceBrandStats = &agentv1.DeviceStats{
-				States: make(map[string]int32),
-			}
-			stats.BrandStats[deviceBrand] = deviceBrandStats
-		}
-		deviceBrandStats.Total++
-		deviceTypeStats.Total++
-
-		if !slot.Enabled {
-			deviceBrandStats.Disabled++
-			deviceTypeStats.Disabled++
-		}
-		if slot.Draining {
-			deviceBrandStats.Draining++
-			deviceTypeStats.Draining++
-		}
-		if slot.Container != nil {
-			deviceBrandStats.States[slot.Container.State.String()]++
-			deviceTypeStats.States[slot.Container.State.String()]++
-		}
-	}
-	return &stats
-}
 
 func (a *apiServer) GetAgents(
 	ctx context.Context, req *apiv1.GetAgentsRequest,
@@ -92,7 +43,6 @@ func (a *apiServer) GetAgents(
 
 	// PERF: can perhaps be done before RBAC.
 	for _, agent := range resp.Agents {
-		agent.SlotStats = SummarizeSlots(agent.Slots)
 		if req.ExcludeSlots {
 			agent.Slots = nil
 		}
